Add support for extra imports in ComponentParser
Added a new feature in ComponentParser so that additional import statements can now be included when parsing components to a file. Extra imports are now added at the beginning of the file. This feature is useful when certain imports are needed for the components defined in the file. Moreover, ComponentParser methods have been changed from class methods to instance methods.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>